+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Fixed
 
+- Apply single band settings to project default layer upon project create [#5454](https://github.com/raster-foundry/raster-foundry/pull/5454)
+
 ### Security
 
 ## [1.47.0](https://github.com/raster-foundry/raster-foundry/compare/1.46.1...1.47.0)
@@ -32,11 +34,13 @@
 ## [1.46.0](https://github.com/raster-foundry/raster-foundry/compare/1.45.0...1.46.0)
 
 ### Fixed
+
 - Fixed tile rendering by updating GeoTrellis to 3.4.1 [#5449](https://github.com/raster-foundry/raster-foundry/pull/5449)
 
 ## [1.45.0(https://github.com/raster-foundry/raster-foundry/compare/1.44.2...1.45.0)
 
 ### Added
+
 - Add support for campaign resource links [#5445](https://github.com/raster-foundry/raster-foundry/pull/5445)
 
 ### Changed
@@ -48,7 +52,6 @@
 
 - Copied back labels should be associated with label classes [#5448](https://github.com/raster-foundry/raster-foundry/pull/5448)
 - `load_development_data --create` should correctly provide a time format parameter [#5448](https://github.com/raster-foundry/raster-foundry/pull/5448)
-
 
 ## [1.44.2](https://github.com/raster-foundry/raster-foundry/compare/1.44.1...1.44.2)
 
@@ -73,7 +76,6 @@
 ### Fixed
 
 - Fix Maven Central release pipeline hanging behavior [#5438](https://github.com/raster-foundry/raster-foundry/pull/5438)
-
 
 ## [1.43.0](https://github.com/raster-foundry/raster-foundry/compare/1.42.0...1.43.0)